pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cg-rera-extractor"
version = "0.1.0"
description = "Extraction framework for the Chhattisgarh RERA portal"
authors = [{ name = "CG RERA Data Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra"
